ArcGIS overhaul
