+++ src/templates/arcgis.html
@@ -17,9 +17,8 @@
 
     <!-- Our CSS and JS -->
     <link rel="stylesheet" href="{{url_for('static', filename='style.css')}}">
-    <script src="{{url_for('static', filename='loadmapendtoend.js')}}"> </script>
-    <!-- <script src="loadmapendtoend.js"></script> -->
-    <!-- style.css -->
+    <script src="{{url_for('static', filename='loadmap.js')}}" type="module"></script>
+    <script src="{{url_for('static', filename='arcgis.js')}}" type="module"></script>
 
     <!-- Favicon -->
     <link rel="shortcut icon" type="image/x-icon" href="{{url_for('static', filename='favicon.ico')}}"/>
